Finished implementing initial conditions parsing.
Also broke parser out into its own file.
