+++ src/CMakeLists.txt
@@ -1,5 +1,6 @@
 # Main library -- we list all source files here (incl. those in subdirectories)
 add_library(rdycore
+  parse_config_file.c
   rdycore.c
   rdyprintf.c
   rdymesh.c
